Script output tile view
Script outputs are now shown on tile view (thumbnail-panels). The console
and file listings can be collapsed (or uncollapsed) as needed. The goal is
for the panels to support self-organising layout with dimensions/aspect ratio
to fit the content (e.g. 2x1 image will be given a different space to 1x1).